refactor(app): split root layout components
