+++ apps/app/src/routes/__root.tsx
@@ -1,19 +1,13 @@
 import type { QueryClient } from "@tanstack/react-query";
-import { createRootRouteWithContext, Outlet, useMatch, useNavigate } from "@tanstack/react-router";
-import {
-  SidebarInset,
-  SidebarProvider,
-  SidebarTrigger,
-  useSidebar,
-} from "@vibest/ui/components/sidebar";
-import { cn } from "@vibest/ui/lib/utils";
+import { createRootRouteWithContext, useMatch, useNavigate } from "@tanstack/react-router";
+import { SidebarProvider } from "@vibest/ui/components/sidebar";
 
 import { AppSidebar } from "@/components/layout/app-sidebar";
-import { usePanelSnapshot } from "@/components/layout/content-panel/react/hooks";
+import { CardPanel } from "@/components/layout/card-panel";
 import { ContentPanelOutlet } from "@/components/layout/content-panel/react/outlet";
 import { ContentPanelProvider } from "@/components/layout/content-panel/react/provider";
 import { RegisterPanels } from "@/components/layout/content-panel/react/register";
-import { ContentPanelToggle } from "@/components/layout/content-panel/react/toggle";
+import { ShellToggle } from "@/components/layout/shell-toggle";
 import { contentPanel, STATIC_PANELS } from "@/content-panel";
 import { useSessionListSync } from "@/features/projects/use-session-list-sync";
 import type { AppClients } from "@/lib/orpc";
@@ -77,88 +71,5 @@ function RootLayout() {
       </ContentPanelProvider>
       <ShellToggle hasTrafficLights={os === "macos"} />
     </SidebarProvider>
-  );
-}
-
-/**
- * Still one fixed toggle for every state rather than a copy inside the sidebar
- * and a copy outside it: the offcanvas sidebar carries an inside toggle
- * off-screen on collapse, and swapping two copies flickers. Only its x moves,
- * animated in step with the sidebar slide.
- *
- * On mobile it is centered in the header's reserved leading slot. On desktop,
- * expanded sits at the sidebar's inner right edge — `--sidebar-width` less the
- * sidebar's own p-1.5, the group's p-2, and the size-7 button. Collapsed takes
- * the corner over, unless macOS's traffic lights already own it.
- */
-function ShellToggle({ hasTrafficLights }: { hasTrafficLights: boolean }) {
-  const { state, isMobile } = useSidebar();
-  const expanded = state === "expanded";
-
-  return (
-    <SidebarTrigger
-      className={cn(
-        "fixed z-30 transition-[left] duration-200 ease-linear [-webkit-app-region:no-drag]",
-        isMobile
-          ? "top-5 left-7 -translate-x-1/2 -translate-y-1/2"
-          : cn(
-              "top-[11px]",
-              expanded
-                ? "left-[calc(var(--sidebar-width)-3.375rem)]"
-                : hasTrafficLights
-                  ? "left-22"
-                  : "left-2",
-            ),
-      )}
-    />
-  );
-}
-
-// Split out so it can read sidebar state via useSidebar().
-function CardPanel() {
-  const { state, isMobile } = useSidebar();
-  // The fixed toggle sits over this header. Reserve its coarse-pointer hit
-  // target on mobile; collapsed desktop also has to clear the traffic lights.
-  const collapsedDesktop = !isMobile && state === "collapsed";
-  // Maximizing squeezes this card to nothing rather than unmounting it: the
-  // route lives inside, and unmounting would dispose the composer's editor.
-  const maximized = usePanelSnapshot((snapshot) => snapshot.presentation === "maximized");
-
-  return (
-    <SidebarInset
-      className={cn(
-        "flex min-h-0 flex-col overflow-hidden border [-webkit-app-region:no-drag] md:peer-data-[variant=inset]:m-1.5 md:peer-data-[variant=inset]:ms-0 md:peer-data-[variant=inset]:peer-data-[state=collapsed]:ms-1.5",
-        // Border and rounding off too — at zero width they would draw a sliver.
-        maximized && "w-0 flex-none border-0 md:peer-data-[variant=inset]:rounded-none",
-      )}
-    >
-      <header
-        className={cn(
-          // Divider is a box-shadow (no layout space) so it can't nudge the
-          // title off the light line; transition animates the collapse-time
-          // padding shift in sync with the sidebar slide.
-          "flex h-10 shrink-0 items-center gap-2 px-4 shadow-[inset_0_-1px_0_var(--color-border)] transition-[padding] duration-200 ease-linear [-webkit-app-region:drag]",
-          isMobile && "ps-14",
-          collapsedDesktop && "ps-30",
-        )}
-      >
-        <div className="flex items-center gap-2 text-sm">
-          <span className="font-medium">New chat</span>
-          <span className="text-muted-foreground">Playground</span>
-        </div>
-        <ContentPanelToggle className="ms-auto [-webkit-app-region:no-drag]" />
-      </header>
-      {/*
-       * Always the Outlet, never a router-state-driven swap: `isLoading` flips
-       * on *every* navigation, including a same-route search-param change like
-       * /draft?projectId=…, and swapping the Outlet out unmounts the active
-       * route — which would dispose the draft composer's editor and drop
-       * whatever the user had typed. Slow route loaders are already covered by
-       * the router's own `defaultPendingComponent` (see router.tsx).
-       */}
-      <div className="flex min-h-0 min-w-0 flex-1 flex-col overflow-hidden">
-        <Outlet />
-      </div>
-    </SidebarInset>
   );
 }
